requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
absl-py==0.7.0            # via tensorflow
astor==0.7.1              # via tensorflow
certifi==2018.11.29       # via requests
chardet==3.0.4            # via requests
click==7.0                # via pip-tools
cycler==0.10.0            # via matplotlib
gast==0.2.2               # via tensorflow
grpcio==1.18.0            # via tensorboard, tensorflow
h5py==2.9.0               # via keras, keras-applications
idna==2.8                 # via requests
keras-applications==1.0.6  # via keras, tensorflow
keras-preprocessing==1.0.5  # via keras, tensorflow
keras==2.2.4
kiwisolver==1.0.1         # via matplotlib
markdown==3.0.1           # via tensorboard
matplotlib==3.0.2
numpy==1.16.0             # via h5py, keras, keras-applications, keras-preprocessing, matplotlib, pandas, scikit-learn, scipy, tensorboard, tensorflow
pandas==0.23.4
pillow==5.4.1             # via python-resize-image
pip-tools==3.2.0
protobuf==3.6.1           # via tensorboard, tensorflow
pyparsing==2.3.1          # via matplotlib
python-dateutil==2.7.5    # via matplotlib, pandas
python-resize-image==1.1.18
pytz==2018.9              # via pandas
pyyaml==3.13              # via keras
requests==2.21.0          # via python-resize-image
scikit-learn==0.20.2
scipy==1.2.0              # via keras, scikit-learn
six==1.12.0               # via absl-py, cycler, grpcio, h5py, keras, keras-preprocessing, pip-tools, protobuf, python-dateutil, tensorboard, tensorflow
tensorboard==1.12.2       # via tensorflow
tensorflow==1.12.0
termcolor==1.1.0          # via tensorflow
urllib3==1.24.1           # via requests
werkzeug==0.14.1          # via tensorboard
wheel==0.32.3             # via tensorboard, tensorflow
